pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-video-translator"
version = "0.1.0"
description = "Pipeline otomasi Python untuk menerjemahkan video analisis catur berbahasa Indonesia menjadi video dengan audio bahasa Inggris dan subtitle burn-in."
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=20.0",
    "faster-whisper",
    "deep-translator",
    "edge-tts",
    "pysrt",
    "google-api-python-client",
    "google-auth",
    "pyrubberband>=0.4.0",
    "soundfile>=0.12.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis",
    "pytest",
    "pytest-asyncio",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
